requirements.txt
streamlit==1.13.0
tensorflow==2.9.1
numpy==1.21.5
pillow==9.0.1
